requirements.txt
matplotlib==3.1.1
scipy==1.1.0
torchvision==0.5.0
numpy==1.17.2
torch==1.4.0
Pillow==7.0.0
